home edited

+++ app/src/main/java/com/darasoylu/shopappcompose/data/model/CategoryModel.kt
@@ -3,6 +3,6 @@ package com.darasoylu.shopappcompose.data.model
 data class CategoryModel(
     val categoryId: String?= null,
     val name: String?= null,
-    val count: Long?= null,
-    val sort: Long?= null
+    val count: Double?= null,
+    val sort: Double?= null
 )
