+++ app/src/main/java/com/darasoylu/shopappcompose/data/model/ProductModel.kt
@@ -4,6 +4,6 @@ data class ProductModel(
     val productId: String?= null,
     val name: String?= null,
     val image: String?= null,
-    val price: Long?= null,
+    val price: Double?= null,
     val categoryId: String?= null
 )
